requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
boto3==1.35.63
botocore==1.35.63
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
dataclasses-json==0.6.7
distro==1.9.0
ecdsa==0.19.0
elastic-transport==8.15.1
elasticsearch==8.16.0
emoji==2.14.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
fastapi==0.115.5
filetype==1.2.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langdetect==1.0.9
langsmith==0.1.143
lxml==5.3.0
marshmallow==3.22.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.24.4
openai==1.55.0
orjson==3.10.11
packaging==24.2
passlib==1.7.4
propcache==0.2.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pypdf==5.1.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.17
PyYAML==6.0.2
rapidfuzz==3.9.7
redis==5.2.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.2
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.67.0
typing-inspect==0.9.0
typing_extensions==4.12.2
unstructured==0.11.8
unstructured-client==0.27.0
urllib3==1.26.20
uvicorn==0.32.0
wrapt==1.16.0
yarl==1.15.2
